# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Componode API
  version: 1.0.0
  description: |
    Componode is an open-source, self-hosted Digital Product Asset Management
    (DPAM) tool. This document describes the REST API exposed by the backend
    (`packages/backend`).

    ## Authentication
    Most endpoints require an authenticated session. Sessions are established
    via `POST /api/v1/auth/login` (or OIDC), which sets an `HttpOnly` session
    cookie (`componode_session`). Mutating requests must also send the CSRF
    token returned at login via the `X-CSRF-Token` header (double-submit
    pattern, cookie `componode_csrf`).

    Public endpoints: `POST /auth/login`, `POST /auth/register`,
    `POST /auth/password/reset/confirm`, the OIDC endpoints, and `GET /metrics`
    (no `/api/v1` prefix; network-restricted in production).

    ## Errors
    Error responses use a consistent envelope:
    `{ "code": "<ERROR_CODE>", "message": "...", "details": <optional> }`.
    Error codes are a controlled enum defined in `packages/core`
    (`ERROR_CODES`).

    ## Conventions
    - All routes are prefixed with `/api/v1` except `GET /metrics`.
    - `GET`/`HEAD` routes are side-effect-free on domain state.
    - List queries exclude `RETIRED` records and `GONE` instances unless
      explicitly requested.
    - Entity IDs are UUID v7; human-readable references use `slug`.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Local username/password login
      description: Public. Rate-limited to 5 requests/minute. Sets session and CSRF cookies.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  pattern: '^[a-z0-9_-]+$'
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/Person' }
                  csrfToken: { type: string }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Self-registration
      description: |
        Public. Only available when `allow_self_registration` is enabled in app
        settings (returns 404 otherwise). Rate-limited to 3 requests/minute.
        Creates a `VIEWER` user and auto-logs in (sets session + CSRF cookies).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  pattern: '^[a-z0-9_-]+$'
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                displayName: { type: string, maxLength: 100 }
      responses:
        '201':
          description: User created and logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/Person' }
                  csrfToken: { type: string }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Username already taken (`AUTH_USERNAME_TAKEN`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/session:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/Person' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password/change:
    post:
      tags: [Auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, minLength: 8, maxLength: 128 }
                newPassword: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '204': { description: Password changed }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Generate a password reset token for a user
      x-permission: "password:reset:generate"
      description: Requires the `password:reset:generate` role (Admin).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
      responses:
        '200':
          description: Reset token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /auth/password/reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm a password reset with a token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string, minLength: 1 }
                newPassword: { type: string, minLength: 8, maxLength: 128 }
      responses:
        '204': { description: Password reset }
        '400':
          description: Validation failed or invalid/expired/used token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /auth/oidc/login:
    post:
      tags: [Auth]
      summary: Initiate OIDC login (redirects to the IdP, PKCE)
      security: []
      parameters:
        - in: query
          name: redirect_uri
          schema: { type: string, default: '/' }
          description: Post-login redirect target
      responses:
        '302': { description: Redirect to the identity provider }
        '503':
          description: OIDC not configured (`OIDC_NOT_CONFIGURED`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/oidc/callback:
    get:
      tags: [Auth]
      summary: OIDC callback — exchanges code, creates session (JIT provisioning), redirects
      security: []
      parameters:
        - in: query
          name: code
          schema: { type: string }
        - in: query
          name: state
          schema: { type: string }
        - in: query
          name: error
          schema: { type: string }
      responses:
        '302':
          description: Redirect to the app (session cookie set; `csrf` query param carries the CSRF token)
        '400':
          description: Missing code/state or OIDC failure
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ----------------------------------------------------------------- Users
...
components:
...
  responses:
    ValidationFailed:
      description: Input validation failed (`VALIDATION_FAILED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Not authenticated (`AUTH_NO_SESSION`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient role (`AUTH_FORBIDDEN`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (`NOT_FOUND`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - AUTH_INVALID_CREDENTIALS
            - AUTH_RATE_LIMITED
            - AUTH_NO_SESSION
            - AUTH_FORBIDDEN
            - AUTH_USERNAME_TAKEN
            - AUTH_RESET_TOKEN_INVALID
            - AUTH_RESET_TOKEN_EXPIRED
            - AUTH_RESET_TOKEN_USED
            - OIDC_NOT_CONFIGURED
            - OIDC_INVALID_STATE
            - OIDC_INVALID_CODE
            - OIDC_TOKEN_VERIFICATION_FAILED
            - OIDC_DISCOVERY_FAILED
            - CSRF_TOKEN_MISMATCH
            - VALIDATION_FAILED
            - NOT_FOUND
            - CONFLICT
            - CYCLE_DETECTED
            - INVALID_EDGE_TYPE
            - REFERENCED
            - TYPE_CHANGE_BLOCKED
            - SLUG_TAKEN
            - SLUG_CONFLICT
            - CONFIG_NOT_FOUND
            - RUN_IN_PROGRESS
            - RUN_NOT_ACTIVE
            - INTERNAL_ERROR
        message: { type: string }
        details: { description: Optional details (e.g. Zod issues) }
...
    Role:
      type: string
      enum: [ADMIN, EDITOR, VIEWER]
...
    Person:
      type: object
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        oidcSubject: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/Role' }
        displayName: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        teamId: { type: string, format: uuid, nullable: true }
        slug: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: componode_session
      description: Session cookie set by login/register/OIDC callback
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on mutating requests (double-submit CSRF)
security:
  - sessionCookie: []